requirements.txt
pandas
pytube
azure-identity
azure-keyvault-secrets
azure-cognitiveservices-speech
srt
pytubefix
moviepy
openai
pydub
opencv-python
azure-functions
